% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accent_remove.R
\name{accent_remove}
\alias{accent_remove}
\title{Remove Vietnamese accent}
\usage{
accent_remove(str)
}
\arguments{
\item{str}{String or vector of string you want to remove accent}
}
\value{
String vector
}
\description{
Remove Vietnamese accent
}
\examples{
accent_remove('tôi muốn ăn cơm')


}
