% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_echart_district.R
\name{custom_echart_district}
\alias{custom_echart_district}
\title{Get json map for echart4r on the district level}
\usage{
custom_echart_district(country = "Vietnam", district)
}
\arguments{
\item{country}{Choose the country, default is Vietnam}

\item{district}{character vector, list all the district want to be included in the map}
}
\description{
Create echart map from GADM map
}
