% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_echart_town.R
\name{custom_echart_town}
\alias{custom_echart_town}
\title{Get json map for echart4r on the town level}
\usage{
custom_echart_town(country = "Vietnam", town)
}
\arguments{
\item{country}{Choose the country, default is Vietnam}

\item{town}{character vector, list all the town want to be included in the map}
}
\description{
Create echart map from GADM map
}
