% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst-helper.R
\name{recode_with}
\alias{recode_with}
\title{Recode vector using multiple pattern/ replacement}
\usage{
recode_with(x, pattern, replace)
}
\arguments{
\item{x}{character vector input}

\item{pattern}{vector of pattern to replace}

\item{replace}{vector of value to replace the corresponding pattern above}
}
\value{
character vector
}
\description{
Recode vector using multiple pattern/ replacement
}
\examples{
data = sample(letters, 20, replace = TRUE)
recode_with(data, c('a','b','c'), c(1,2,3))
}
