% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_vnsf.R
\docType{data}
\name{small_vnsf}
\alias{small_vnsf}
\title{Small sf data of Vietnam map}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 63 rows and 2 columns.
}
\usage{
small_vnsf
}
\description{
Hoang Sa and Truong Sa are not included
}
\examples{
# This sf map data is easy to visualize with ggplot2
library(ggplot2)
library(magrittr)

ggplot(small_vnsf, aes(geometry = geometry))+
geom_sf()+
coord_sf()

}
\keyword{datasets}
