% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_to_json.R
\name{sp_to_json}
\alias{sp_to_json}
\title{Transform sp to json}
\usage{
sp_to_json(sp = NULL, scope = NULL)
}
\arguments{
\item{sp}{sp object}

\item{scope}{the scope of map data: town or district. \cr
Eample of town: Ha Noi, Ha Tinh. \cr
Example of district: Vinh, Hong Linh, Nghi Xuan}
}
\value{
a json map ready to feed echarts4r::e_map_register()
}
\description{
Get json map for echart4r from sp data download from GADM (raster)
}
