% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_remove_accent.R
\name{str_remove_accent}
\alias{str_remove_accent}
\title{Remove accent ingeneral}
\usage{
str_remove_accent(str)
}
\arguments{
\item{str}{String or vector of string you want to remove accent}
}
\value{
String vector
}
\description{
Wrap around stringi
}
\examples{
str_remove_accent('tôi muốn ăn cơm')

}
