% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyst-helper.R
\name{top_hilo}
\alias{top_hilo}
\title{Indicate top high, low of numeric vector}
\usage{
top_hilo(x, n = 10)
}
\arguments{
\item{x}{input vector}

\item{n}{number of top value each size}
}
\value{
numeric vecotr of 0, -1, and 1
}
\description{
Return vector of 1,-1,0 indicate top high, low of input vector
1 for top high, -1 for top low, 0 for the others
This function work on conjunction with geom_col(), to show for example top n high and lowest point
of some value
}
