% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntown.R
\docType{data}
\name{vntown}
\alias{vntown}
\title{Geojson data map of Vietnam}
\format{
An object of class \code{geo_list} of length 3.
}
\usage{
vntown
}
\description{
(Hoang Sa and Truong Sa are not included)
}
\examples{
# To use this geojson map with echart4r package
library(echarts4r)
library(magrittr)
data.frame(province = 'Ha Tinh', value = 1) \%>\%
e_chart(province) \%>\%
e_map_register('vn', vntown) \%>\%
e_map(value, map = 'vn') \%>\%
e_visual_map(value)
}
\keyword{datasets}
