% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_echart_town.R
\name{custom_echart_town}
\alias{custom_echart_town}
\title{Create Echart Map From GADM Map}
\usage{
custom_echart_town(country = "Vietnam", town)
}
\arguments{
\item{country}{Choose the country, default is Vietnam}

\item{town}{character vector, list all the town want to be included in the map}
}
\value{
a json map ready to feed echarts4r::e_map_register()
}
\description{
Get json map for echart4r on the town level
}
