% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-cafeF.R
\name{get_cafeF}
\alias{get_cafeF}
\title{Get stock data from cafef.vn}
\usage{
get_cafeF(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{the code of stock}

\item{start_date}{the start date}

\item{end_date}{the end date}
}
\value{
a tibble contain all data of the stock in that period
}
\description{
Get stock data from cafef.vn
}
\examples{
get_cafeF('ACB', "1/1/2020", "1/2/2020")
}
