% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vndirect.R
\name{get_vndirect}
\alias{get_vndirect}
\title{Get stock data from vndirect}
\usage{
get_vndirect(symbol, start_date = NULL, end_date = NULL)
}
\arguments{
\item{symbol}{the symbol of stock}

\item{start_date}{the start date}

\item{end_date}{the end_date}
}
\description{
Get stock data from vndirect
}
