% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_to_json.R
\name{sp_to_json}
\alias{sp_to_json}
\title{Transform sp to json}
\usage{
sp_to_json(sp = NULL, scope = NULL, std_name = TRUE)
}
\arguments{
\item{sp}{Sp object. Might be downloaded from raster:getdata()}

\item{scope}{The scope of map data: town or district. \cr
Eample of town: Ha Noi, Ha Tinh. \cr
Example of district: Vinh, Hong Linh, Nghi Xuan}

\item{std_name}{Should use standard name  or not \cr
For example, "Hà Tĩnh" is non-standard name, while "Ha Tinh" is standard name}
}
\value{
A json map ready to feed echarts4r::e_map_register()
}
\description{
Get json map for echart4r from sp data download from GADM (raster)
}
